+++ backend/app/main.py
@@ -1,10 +1,11 @@
 import os
 import logging
 from dotenv import load_dotenv, find_dotenv
-from fastapi import FastAPI, BackgroundTasks
+from fastapi import FastAPI, Request, BackgroundTasks
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.staticfiles import StaticFiles
 from fastapi.templating import Jinja2Templates
+from fastapi.responses import HTMLResponse
 from .routers.database import admin, schools, calendar, timetable, curriculum, department, teacher, student
 from .routers.transcribe import whisper_live
 from .routers.llm import ollama, openai
@@ -24,8 +25,11 @@ app.add_middleware(
     expose_headers=["*"]
 )
 
-app.mount("/static", StaticFiles(directory="fastapi_frontend/static"), name="static")
-templates = Jinja2Templates(directory="fastapi_frontend/templates")
+# Get the absolute path to the fastapi_frontend directory
+frontend_directory = os.path.join(os.path.dirname(__file__), '../../fastapi_frontend')
+
+app.mount("/static", StaticFiles(directory=os.path.join(frontend_directory, 'static')), name="static")
+templates = Jinja2Templates(directory=os.path.join(frontend_directory, 'templates'))
 
 # Database Routes
 app.include_router(admin.router, prefix="/database/admin", tags=["Admin"])
@@ -43,3 +47,7 @@ app.include_router(whisper_live.router, prefix="/transcribe/live", tags=["Transc
 # LLM Routes
 app.include_router(ollama.router, prefix="/llm", tags=["LLM"])
 app.include_router(openai.router, prefix="/llm", tags=["LLM"])
+
+@app.get("/dashboard", response_class=HTMLResponse)
+async def get_dashboard(request: Request):
+    return templates.TemplateResponse("dashboard.html", {"request": request})
