+++ backend/app/routers/database/admin.py
@@ -1,10 +1,10 @@
 from fastapi import APIRouter, Depends
 from app.dependencies import admin_dependency
 
-import app.routers.database.tools.neo4j_driver_tools as driver
-import app.routers.database.tools.neo4j_session_tools as session
-import app.routers.database.tools.neo4j_http_tools as http
-import app.routers.database.tools.query_library as query
+import app.modules.database.tools.neo4j_driver_tools as driver
+import app.modules.database.tools.neo4j_session_tools as session
+import app.modules.database.tools.neo4j_http_tools as http
+import app.modules.database.tools.query_library as query
 
 router = APIRouter()
