+++ backend/app/routers/database/curriculum.py
@@ -1,7 +1,8 @@
 from fastapi import APIRouter, Depends, File, UploadFile
 from ...dependencies import admin_dependency
-from ...modules.driver_tools import get_driver, create_node, create_relationship
-from ...modules.get_planner import get_excel_sheets
+from app.modules.database.tools.neo4j_driver_tools import get_driver
+from app.modules.database.tools.neo4j_http_tools import create_node, create_relationship
+from app.modules.database.tools.xl_planner_tools import get_excel_sheets
 
 router = APIRouter()
