+++ backend/app/routers/transcribe/whisper_live.py
@@ -1,7 +1,6 @@
-from fastapi import APIRouter, BackgroundTasks, WebSocket
-from ...modules.whisperlive_client_new import client
+from fastapi import APIRouter, WebSocket
+from app.modules.whisperlive_client import Client
 import os
-import time
 
 router = APIRouter()
 
@@ -19,7 +18,7 @@ async def websocket_endpoint(websocket: WebSocket):
     user_transcript_dir = setup_directories(user_dir, user_id)
     
     await websocket.accept()
-    transcription_client = client.TranscriptionClient(
+    transcription_client = Client.TranscriptionClient(
         websocket=websocket,
         lang="en",
         translate=False,
@@ -30,28 +29,3 @@ async def websocket_endpoint(websocket: WebSocket):
     )
     await transcription_client.handle_connection()  # Adjusted method to handle the connection
     await websocket.close()
-
-
-from fastapi import APIRouter, WebSocket
-from modules.whisperlive_client_new import Client as TranscriptionClient
-
-router = APIRouter()
-
-@router.websocket("/whisper-live-audio-stream")
-async def websocket_endpoint(websocket: WebSocket):
-    await websocket.accept()
-    try:
-        # Assuming the server details and other configurations are set as environment variables or similar
-        client = TranscriptionClient(
-            ws=websocket,
-            host="localhost",  # Backend server host
-            port=9090,         # Backend server port
-            lang="en",
-            translate=False,
-            use_vad=True,
-            model="small",
-            srt_file_path="path_to_srt_file.srt"  # Update with actual path
-        )
-        await client.handle_connection()  # This method should be implemented to manage the connection lifecycle
-    finally:
-        await websocket.close()
