+++ fastapi_frontend/templates/dashboard.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Admin Dashboard</title>
-    <link rel="stylesheet" href="{{ url_for('static', filename='example.css') }}">
+    <link rel="stylesheet" href="{{ url_for('static', path='example.css') }}">
 </head>
 <body>
     <div class="container">
